+++ src/io/mod.rs
@@ -26,35 +26,6 @@ cfg_if::cfg_if! {
     }
 }
 
-cfg_if::cfg_if! {
-    if #[cfg(feature = "mwalib")] {
-        use std::ops::Range;
-        use mwalib::CorrelatorContext;
-    }
-}
-
-/// The container has visibilities which can be read by passing in a mwalib
-/// context and the range of values to read.
-#[cfg(feature = "mwalib")]
-pub trait VisRead: Sync + Send {
-    /// Read the visibilities and weights for the selected timesteps, coarse
-    /// channels and baselines into the provided arrays.
-    ///
-    /// # Errors
-    ///
-    /// Can throw `IOError` if there is an issue reading.
-    ///
-    fn read_vis_mwalib(
-        &self,
-        jones_array: ArrayViewMut3<Jones<f32>>,
-        weight_array: ArrayViewMut3<f32>,
-        context: &CorrelatorContext,
-        timestep_range: &Range<usize>,
-        coarse_chan_range: &Range<usize>,
-        baseline_idxs: &[usize],
-    ) -> Result<(), IOError>;
-}
-
 /// The container can accept a chunk of visibilities to be written.
 pub trait VisWrite {
     /// Write a chunk of visibilities, contextualised with a [`VisContext`].
@@ -71,14 +42,11 @@ pub trait VisWrite {
     ///
     /// `vis_ctx` - a [`VisContext`] which contextualises each axis of the visibilities.
     ///
-    /// `draw_progress` - whether or not to draw a progress bar.
-    ///
     fn write_vis(
         &mut self,
         vis: ArrayView3<Jones<f32>>,
         weights: ArrayView3<f32>,
         vis_ctx: &VisContext,
-        draw_progress: bool,
     ) -> Result<(), IOError>;
 
     /// When all visibilities have been given to this [`VisWrite`] implementor,